+++ src/model/user.rs
@@ -8,7 +8,7 @@ use crate::service::user::level::LevelInfo;
 #[sea_orm(table_name = "user")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     #[sea_orm(unique)]
     pub username: String,
     #[serde(skip_serializing)]
@@ -19,24 +19,25 @@ pub struct Model {
     pub available: bool,
     #[serde(skip_serializing)]
     pub level: Vec<String>,
-    pub root: i32,
+    pub root: i64,
     pub create_time: DateTime,
     pub update_time: DateTime,
 }
 
 #[derive(Serialize)]
 pub struct UserExtended {
-    pub id: i32,
+    pub id: i64,
     pub username: String,
     pub phone: String,
     pub email: Option<String>,
     pub available: bool,
     pub level: LevelInfo,
-    pub root: i32,
+    pub root: i64,
     pub used_space: f64,
     pub create_time: DateTime,
     pub update_time: DateTime,
 }
+
 
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
